+++ src/Component/Navbar.jsx
@@ -1,6 +1,6 @@
 function NavBar(){
     return (
-        <nav className="navbar flex items-center space-x-4 border-b-2">
+        <nav className="navbar flex items-center space-x-4 border-b-2 color-black">
             <Link to = "/analysis">
                 <img src="https://cdn-icons-png.flaticon.com/512/1570/1570889.png" alt="analysis" />
             </Link>
